fix: abstract duplicated code
feat(assignables): new hook usePendingEvaluationsCount
